refactor(queries): consolidate placeholder steps into a single file
