+++ src/queries/index.ts
@@ -15,7 +15,6 @@ export * from './link';
 export * from './links';
 export * from './name';
 export * from './placeholder';
-export * from './placeholders';
 export * from './testid';
 export * from './text';
 export * from './title';
